# openapi-spec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "San Pablo API",
    "version": "1.0.0",
    "description": "API for art store/bookstore"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "OrderItem": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "productId": { "type": "string", "format": "uuid", "nullable": true },
          "productSku": { "type": "string" },
          "productName": { "type": "string" },
          "quantity": { "type": "integer" },
          "unitPrice": { "type": "string" },
          "totalPrice": { "type": "string" }
        }
      },
      "Order": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "orderNumber": { "type": "string", "example": "SP-20240115-XYZ1" },
          "userId": { "type": "string", "format": "uuid", "nullable": true },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "confirmed",
              "processing",
              "shipped",
              "delivered",
              "cancelled"
            ]
          },
          "contactFullName": { "type": "string" },
          "contactEmail": { "type": "string" },
          "contactPhone": { "type": "string" },
          "contactAddress": { "type": "string" },
          "subtotal": { "type": "string" },
          "total": { "type": "string" },
          "customerNotes": { "type": "string", "nullable": true },
          "internalNotes": { "type": "string", "nullable": true },
          "items": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/OrderItem" }
          },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
...
      "Product": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "sku": { "type": "string" },
          "name": { "type": "string" },
          "description": { "type": "string" },
          "regularPrice": { "type": "string" },
          "salePrice": { "type": "string", "nullable": true },
          "specialPrice": { "type": "string", "nullable": true },
          "imageUrl": { "type": "string", "nullable": true },
          "thumbnailUrl": { "type": "string", "nullable": true },
          "videoUrl": { "type": "string", "nullable": true },
          "weight": { "type": "string", "nullable": true },
          "dimensionLength": { "type": "string", "nullable": true },
          "dimensionWidth": { "type": "string", "nullable": true },
          "dimensionHeight": { "type": "string", "nullable": true },
          "categoryId": { "type": "string", "format": "uuid", "nullable": true },
          "isActive": { "type": "boolean" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "PaginationMeta": {
        "type": "object",
        "properties": {
          "page": { "type": "integer" },
          "limit": { "type": "integer" },
          "total": { "type": "integer" },
          "totalPages": { "type": "integer" }
        }
      },
...
    }
  },
...
  "paths": {
...
    "/api/products/{id}": {
      "get": {
        "tags": ["Products"],
        "summary": "Get product by ID",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": { "description": "Product found" },
          "404": { "description": "Product not found" }
        }
      },
      "put": {
        "tags": ["Products"],
        "summary": "Update a product",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/Product" }
            }
          }
        },
        "responses": {
          "200": { "description": "Product updated" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" },
          "404": { "description": "Product not found" }
        }
      },
      "delete": {
        "tags": ["Products"],
        "summary": "Delete a product (soft delete)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "204": { "description": "Product deleted" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" },
          "404": { "description": "Product not found" }
        }
      }
    },
...
    "/api/orders": {
      "get": {
        "tags": ["Orders"],
        "summary": "List orders",
        "description": "Users see their own orders, admins see all",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": { "type": "integer", "default": 1 }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": { "type": "integer", "default": 20 }
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string",
              "enum": [
                "pending",
                "confirmed",
                "processing",
                "shipped",
                "delivered",
                "cancelled"
              ]
            }
          },
          {
            "in": "query",
            "name": "userId",
            "schema": { "type": "string", "format": "uuid" },
            "description": "Filter by user ID (admin only)"
          }
        ],
        "responses": {
          "200": {
            "description": "Orders list with pagination",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Order" }
                    },
                    "meta": { "$ref": "#/components/schemas/PaginationMeta" }
                  }
                }
              }
            }
          },
          "401": { "description": "Authentication required" }
        }
      }
    },
    "/api/orders/{id}": {
      "get": {
        "tags": ["Orders"],
        "summary": "Get order by ID",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": { "description": "Order details" },
          "403": { "description": "Access denied" },
          "404": { "description": "Order not found" }
        }
      },
      "delete": {
        "tags": ["Orders"],
        "summary": "Delete order (soft delete, admin only)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "204": { "description": "Order deleted" },
          "403": { "description": "Admin access required" },
          "404": { "description": "Order not found" }
        }
      }
    },
    "/api/orders/number/{orderNumber}": {
      "get": {
        "tags": ["Orders"],
        "summary": "Get order by order number",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "in": "path",
            "name": "orderNumber",
            "required": true,
            "schema": { "type": "string" },
            "example": "SP-20240115-XYZ1"
          }
        ],
        "responses": {
          "200": { "description": "Order details" },
          "403": { "description": "Access denied" },
          "404": { "description": "Order not found" }
        }
      }
    },
    "/api/orders/{id}/status": {
      "put": {
        "tags": ["Orders"],
        "summary": "Update order status (admin only)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["status"],
                "properties": {
                  "status": {
                    "type": "string",
                    "enum": [
                      "pending",
                      "confirmed",
                      "processing",
                      "shipped",
                      "delivered",
                      "cancelled"
                    ]
                  },
                  "internalNotes": { "type": "string", "maxLength": 2000 }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Order updated" },
          "403": { "description": "Admin access required" },
          "404": { "description": "Order not found" }
        }
      }
    },
...
  }
}